Global: moved all database related code to /db, entry point /db/index.js
