+++ db/index.js
@@ -1,0 +1,7 @@
+const utils = require('./utils');
+const models = require('./models');
+
+module.exports = {
+  utils,
+  models,
+};
