+++ db/models/story/story-instance-methods/index.js
