+++ db/models/story/story-static-methods/index.js
