+++ db/utils/index.js
@@ -1,0 +1,7 @@
+const testUtils = require('./test-utils');
+const getDatabaseURI = require('./get-db-uri');
+
+module.exports = {
+  testUtils,
+  getDatabaseURI,
+};
